requirements.txt
arize-phoenix-otel>=0.6.0
openinference-instrumentation-openai>=0.1.12
openinference-instrumentation-langchain>=0.1.29
openai>=1.35.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
chromadb>=0.5.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
pandas>=2.0.0
rich>=13.0.0
